pjt2/project1/backend/requirements.txt
absl-py==1.4.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
Brotli==1.0.9
bs4==0.0.1
cached-property==1.5.2
cachetools==5.3.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cmake==3.25.0
colorama==0.4.5
commonmark==0.9.1
ConfigArgParse==1.5.3
contourpy==1.0.6
cryptography==37.0.4
cycler==0.11.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
Django==3.2.12
django-cors-headers==3.13.0
django-extensions==3.2.0
django-mathfilters==1.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dlib==19.24.0
dnspython==2.2.1
docopt==0.6.2
entrypoints==0.4
exceptiongroup==1.0.2
executing==0.8.3
ez-setup==0.9
Faker==13.15.1
fastapi==0.88.0
fastjsonschema==2.16.1
Flask==2.2.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
flatbuffers==23.1.21
fonttools==4.38.0
gast==0.4.0
gevent==22.10.2
geventhttpclient==2.0.8
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.1
grpcio==1.51.1
h11==0.13.0
h5py==3.8.0
httptools==0.5.0
idna==3.3
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.7.2
jupyter-client==7.3.4
jupyter-contrib-core==0.4.0
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.11.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.5.0
jupyterlab-pygments==0.2.2
keras==2.11.0
kiwisolver==1.4.4
libclang==15.0.6.1
locust==2.14.0
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.3
mistune==0.8.4
msgpack==1.0.4
mysql-connector-python==8.0.31
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.23.4
oauthlib==3.2.2
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
pipreqs==0.4.11
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.6
psutil==5.9.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybrowsers==0.5.0
pycparser==2.21
pydantic==1.10.2
pygame==2.1.2
Pygments==2.12.0
PyJWT==2.6.0
pymongo==4.3.3
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-slugify==6.1.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==23.2.0
requests==2.28.1
requests-oauthlib==1.3.1
roundrobin==0.0.4
rsa==4.9
seaborn==0.12.1
selenium==3.14.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.3.0
starlette==0.22.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
terminado==0.15.0
text-unidecode==1.3
tinycss2==1.1.1
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.4.0
tzdata==2021.5
tzlocal==4.2
urllib3==1.26.10
uvicorn==0.20.0
watchfiles==0.18.1
wcwidth==0.2.5
webdriver-manager==3.8.0
webencodings==0.5.1
websockets==10.4
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.1.0
yarg==0.1.9
zope.event==4.6
zope.interface==5.5.2
